+++ build.gradle
@@ -112,6 +112,25 @@ project(":core") {
         testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
         testImplementation 'org.junit.jupiter:junit-jupiter-params:5.9.2'
         testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+    }
+
+    test {
+        useJUnitPlatform()
+        finalizedBy jacocoTestReport
+    }
+
+    jacocoTestReport {
+        reports {
+            xml.enabled = true
+        }
+    }
+}
+
+project(":tests") {
+    apply plugin: "java"
+
+    dependencies {
+        implementation project(":core")
 
         implementation "junit:junit:4.13.2"
         implementation "org.mockito:mockito-all:1.10.19"
@@ -125,16 +144,5 @@ project(":core") {
         testImplementation "com.badlogicgames.gdx:gdx-backend-headless:$gdxVersion"
         testImplementation "com.badlogicgames.gdx:gdx:$gdxVersion"
         testImplementation "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
-    }
-
-    test {
-        useJUnitPlatform()
-        finalizedBy jacocoTestReport
-    }
-
-    jacocoTestReport {
-        reports {
-            xml.enabled = true
-        }
     }
 }
