+++ tests/src/it/units/sdm/project/GdxTestRunner.java
@@ -1,16 +1,14 @@
-package runners;
+package it.units.sdm.project;
 
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.backends.headless.HeadlessApplication;
 import com.badlogic.gdx.backends.headless.HeadlessApplicationConfiguration;
 import com.badlogic.gdx.graphics.GL20;
-import org.jetbrains.annotations.NotNull;
 import org.junit.runner.notification.RunNotifier;
 import org.junit.runners.BlockJUnit4ClassRunner;
 import org.junit.runners.model.FrameworkMethod;
 import org.junit.runners.model.InitializationError;
-import org.mockito.Mockito;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -19,7 +17,6 @@ import static org.mockito.Mockito.mock;
 
 public class GdxTestRunner extends BlockJUnit4ClassRunner implements ApplicationListener {
 
-    @NotNull
     private final Map<FrameworkMethod, RunNotifier> invokeInRender = new HashMap<>();
     public GdxTestRunner(Class<?> testClass) throws InitializationError {
         super(testClass);
